Add files via upload
Hello and welcome to my GitHub repository for the Quotes Web Scraping project!

In this project, I have created a Python script to scrape quotes from the website https://quotes.toscrape.com/ and store them in an excel file. The script uses the BeautifulSoup library to parse the HTML content of the website and extract the relevant data.

I have also included a Jupyter Notebook file named Quotes_Web_Scraping.ipynb, which provides some basic analysis of the scraped quotes. It also allow to download the dataset in the local system with file name Quotes_to_Scrape.xlsx.